pyproject.toml
[tool.poetry]
name = "llm-conversion-framework"
version = "2.3.0"
description = "Orchestrator GUI und CLI für LLM Cross-Compilierung"
authors = ["Smilez1985"]
readme = "README.md"
packages = [{include = "orchestrator"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"

# --- 1. Core & Installer Dependencies ---
requests = "^2.31.0"
psutil = "^5.9.0"
pyyaml = "^6.0.1"
packaging = "^23.2"
python-dotenv = "^1.0.0"  # Hinzugefügt wie besprochen

# --- 2. Docker Orchestrierung ---
docker = "^7.0.0"

# --- 3. GUI & CLI ---
pyside6 = "^6.6.0"
click = "^8.1.7"
rich = "^13.7.0"

# --- 4. Security & Utilities ---
cryptography = "^41.0.0"

# --- 5. AI Agent & RAG (Host-Seite) ---
litellm = "^1.0.0"
qdrant-client = "^1.7.0"

# --- 6. Crawler & Ingestion (Host-Seite) ---
beautifulsoup4 = "^4.12.0"
langchain = "^0.1.0"
langchain-community = "^0.0.20"
pypdf = "^4.2.0"
tiktoken = "^0.7.0"
langdetect = "^1.0.9"

# --- 7. Windows Specifics ---
pywin32 = {version = "^306", markers = "sys_platform == 'win32'"}
winshell = {version = "^0.6", markers = "sys_platform == 'win32'"}

# --- 8. Optional Deployment (SSH) ---
paramiko = "^3.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
flake8 = "^7.0.0"
bandit = "^1.7.0"
safety = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
llm-gui = "orchestrator.main:main"
llm-cli = "orchestrator.cli:cli"
